refactor(bridge): extract action census
